% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cytoscapeRegulon}
\alias{cytoscapeRegulon}
\title{CytoscapeRegulon}
\usage{
cytoscapeRegulon(x)
}
\arguments{
\item{x}{Is the folder path of the GENSOR unit}
}
\details{
Creates a cytoscape network from a tab delimited GENSOR unit files.
}
\examples{
cytoscapeRegulon("/Users/daniel/Proyecto-Cytoscape/SRsGUs/LacI")

}
\author{
Antonio Daniel Martinez Gutierrez
}
